requirements.txt
requests>=2.28
beautifulsoup4>=4.11
openai==1.30.5
python-dotenv>=1.0.0